python3 support, remove icu, sdl, angle (save 547MiB)
Remove icu and sdl by just commenting them out in DEPS and changing
skia_use_system_icu to true. Remove angle2 as it is not needed for the
base libskia.a.

Update third_party/third_party.gni to work around a subgn bug. It is not
possible for a parent project to reach in via subgn unless the target is a
static_library. A source_set does not produce a ninja target that the parent
could then have a dependency on.

Add gSkDebugVPrinter for redirecting output to a log.

Update DEPS to more recent versions.